requirements.txt
asgiref==3.8.1
Brotli==1.1.0
Django==5.2.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
sqlparse==0.5.3
whitenoise==6.9.0
